e2e_tests/test_requirements.txt
pytest-playwright==0.3.3
allure-pytest
